server/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
mysqlclient==2.2.7
PyJWT==2.9.0
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.1
